+++ firmware/include/WebApi.h
@@ -2,12 +2,14 @@
 #include <Arduino.h>
 #include "Radio.h"
 #include "Crypto.h"
+#include "WifiConfig.h" 
 
 class WebApi {
 public:
-    void begin(Radio* radio, Crypto* crypto, const char* apSsid, const char* apPassword);
-
+    void begin(Radio* radio, Crypto* crypto, WifiConfig* wifi, const char* apSsid, const char* apPassword);
+    
 private:
     static Radio* radioRef;
     static Crypto* cryptoRef; 
+    static WifiConfig* wifiConfigRef;
 };